+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Huffman code".
 - Tests for "Number of paths between 2 points".
 - Tests for "The lost child.Episode-1".
+- Tests for "Gravity tumbler".
 
 ## [1.12.0] - 2022-09-01
 ### Added